make adjusment code and folder
