+++ package.json
@@ -14,7 +14,7 @@
     "react-dom": "^17.0.2",
     "react-notifications-menu": "^1.0.1",
     "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
+    "react-scripts": "^4.0.3",
     "web-vitals": "^1.1.2"
   },
   "scripts": {
